# artifacts/p10/openapi.json
{
  "components": {
    "schemas": {
      "CreateSessionRequest": {
        "properties": {
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          },
          "extractor_mode": {
            "anyOf": [
              {
                "enum": [
                  "real_llm",
                  "fake",
                  "fallback"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extractor Mode"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "rag_enabled": {
            "default": true,
            "title": "Rag Enabled",
            "type": "boolean"
          },
          "store_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Backend"
          }
        },
        "title": "CreateSessionRequest",
        "type": "object"
      },
      "CreateSessionResponse": {
        "properties": {
          "api_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Version"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "extractor_mode": {
            "enum": [
              "real_llm",
              "fake",
              "fallback"
            ],
            "title": "Extractor Mode",
            "type": "string"
          },
          "rag_enabled": {
            "title": "Rag Enabled",
            "type": "boolean"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "store_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Backend"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "extractor_mode",
          "rag_enabled",
          "created_at",
          "turn_count"
        ],
        "title": "CreateSessionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplayRequest": {
        "properties": {
          "allow_write": {
            "default": false,
            "title": "Allow Write",
            "type": "boolean"
          }
        },
        "title": "ReplayRequest",
        "type": "object"
      },
      "ReplayResponse": {
        "properties": {
          "final_report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Report"
          },
          "final_state": {
            "additionalProperties": true,
            "title": "Final State",
            "type": "object"
          },
          "graph_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Graph Events",
            "type": "array"
          },
          "replay_status": {
            "default": "ok_read_only",
            "title": "Replay Status",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "turns": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ReplayResponse",
        "type": "object"
      },
...
      "SessionReportResponse": {
        "properties": {
          "evidence": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "final_report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Report"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "missing_core_fields": {
            "items": {
              "type": "string"
            },
            "title": "Missing Core Fields",
            "type": "array"
          },
          "next_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "report_available": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Available"
          },
          "risk_reasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Reasons"
          },
          "risk_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Status"
          },
          "safety_disclaimer": {
            "default": "本系统仅用于问诊信息整理和风险提示，不构成诊断或治疗建议，不能替代医生判断。如出现持续高热、胸痛、呼吸困难、便血、意识异常、剧烈腹痛等高风险信号，应及时线下就医。",
            "title": "Safety Disclaimer",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "ready"
        ],
        "title": "SessionReportResponse",
        "type": "object"
      },
      "SessionStateResponse": {
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "missing_core_fields": {
            "items": {
              "type": "string"
            },
            "title": "Missing Core Fields",
            "type": "array"
          },
          "next_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question"
          },
          "risk_flags_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Flags Status"
          },
          "risk_rule_ids": {
            "items": {
              "type": "string"
            },
            "title": "Risk Rule Ids",
            "type": "array"
          },
          "safety_disclaimer": {
            "default": "本系统仅用于问诊信息整理和风险提示，不构成诊断或治疗建议，不能替代医生判断。如出现持续高热、胸痛、呼吸困难、便血、意识异常、剧烈腹痛等高风险信号，应及时线下就医。",
            "title": "Safety Disclaimer",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "session_id",
          "turn_count",
          "state"
        ],
        "title": "SessionStateResponse",
        "type": "object"
      },
...
      "TurnRequest": {
        "properties": {
          "debug": {
            "default": false,
            "title": "Debug",
            "type": "boolean"
          },
          "extractor_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extractor Backend"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "user_input": {
            "minLength": 1,
            "title": "User Input",
            "type": "string"
          }
        },
        "required": [
          "user_input"
        ],
        "title": "TurnRequest",
        "type": "object"
      },
      "TurnResponse": {
        "properties": {
          "event_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Count"
          },
          "fallback_used": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Used"
          },
          "final_report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Report"
          },
          "graph_runtime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Runtime"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "missing_core_fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missing Core Fields"
          },
          "next_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question"
          },
          "retrieved_evidence_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved Evidence Count"
          },
          "risk_flags_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Flags Status"
          },
          "risk_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Risk Reasons",
            "type": "array"
          },
          "risk_rule_ids": {
            "items": {
              "type": "string"
            },
            "title": "Risk Rule Ids",
            "type": "array"
          },
          "risk_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Status"
          },
          "safety_disclaimer": {
            "default": "本系统仅用于问诊信息整理和风险提示，不构成诊断或治疗建议，不能替代医生判断。如出现持续高热、胸痛、呼吸困难、便血、意识异常、剧烈腹痛等高风险信号，应及时线下就医。",
            "title": "Safety Disclaimer",
            "type": "string"
          },
          "safety_rewrite_used": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Safety Rewrite Used"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "state_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Summary"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "turn_count": {
            "title": "Turn Count",
            "type": "integer"
          },
          "turn_id": {
            "title": "Turn Id",
            "type": "string"
          },
          "warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          }
        },
        "required": [
          "session_id",
          "turn_id",
          "turn_count",
          "state"
        ],
        "title": "TurnResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "P1.1 minimal FastAPI wrapper around the existing LangGraph workflow.",
    "title": "TCM-Assistant Minimal API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sessions": {
      "post": {
        "operationId": "create_api_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Api Session"
      }
    },
...
    "/sessions/{session_id}/replay": {
      "post": {
        "operationId": "replay_api_session_sessions__session_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ReplayRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay Api Session"
      }
    },
    "/sessions/{session_id}/report": {
      "get": {
        "operationId": "get_api_session_report_sessions__session_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Api Session Report"
      },
      "post": {
        "operationId": "post_api_session_report_sessions__session_id__report_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Api Session Report"
      }
    },
    "/sessions/{session_id}/state": {
      "get": {
        "operationId": "get_api_session_state_sessions__session_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Api Session State"
      }
    },
...
    "/sessions/{session_id}/turn": {
      "post": {
        "operationId": "submit_turn_sessions__session_id__turn_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Turn"
      }
    },
...
  }
}